+++ src/main/java/com/buuz135/simpleclaims/claim/player_name/PlayerNameTracker.java
@@ -47,6 +47,10 @@ public class PlayerNameTracker {
         names.put(uuid, new PlayerName(uuid, name));
     }
 
+    public HashMap<UUID, PlayerName> getNamesMap() {
+        return names;
+    }
+
     public static class PlayerName{
 
         public static final BuilderCodec<PlayerName> CODEC = BuilderCodec.builder(PlayerName.class, PlayerName::new)